# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EnergieGo API",
        "description": "Documentation OpenAPI pour l'API EnergieGo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/marques": {
            "get": {
                "summary": "Lister les marques",
                "operationId": "03787a04650457a20881200057ea409e",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Marque"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Créer une marque",
                "operationId": "d6742751abad93fc4e5b8b2505c311e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Marque"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Marque"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/marques/{id}": {
            "get": {
                "summary": "Afficher une marque",
                "operationId": "5f11e45542cd15be491207048afb4e89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Marque"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "put": {
                "summary": "Mettre à jour une marque",
                "operationId": "1158dc69899538d010715a26d138829d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Marque"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Marque"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Supprimer une marque",
                "operationId": "dc28b5d662995167a39c7d9ec4372b57",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                }
            }
        },
        "/api/paiements": {
            "get": {
                "summary": "Lister les paiements",
                "operationId": "228c5720fb8a55f052efcfa2e73505e9",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Paiement"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Créer un paiement",
                "operationId": "95b59af066ce7e9cf00b06dbd44dd314",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Paiement"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Paiement"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/paiements/{id}": {
            "get": {
                "summary": "Afficher un paiement",
                "operationId": "af5e37d407a752cdf5fae2d5f2e4ed96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Paiement"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Mettre à jour un paiement",
                "operationId": "1cb40486f51570497e35760aa338048f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Paiement"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Paiement"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Supprimer un paiement",
                "operationId": "8909c66d08ffdac26efd8e22c11022f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                }
            }
        },
        "/api/produits": {
            "get": {
                "summary": "Lister les produits",
                "operationId": "059ba901d56d1b85845397dcc5120bd2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Produit"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Créer un produit",
                "operationId": "1d2c1c7cee2e131dfeb85184f2824a90",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Produit"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Produit"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/produits/{id}": {
            "get": {
                "summary": "Afficher un produit",
                "operationId": "4f89d6d658ffba7d843303b681d52762",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Produit"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "put": {
                "summary": "Mettre à jour un produit",
                "operationId": "05d27b4688f87d159b38ecb2f0d03dc1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Produit"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Produit"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "delete": {
                "summary": "Supprimer un produit",
                "operationId": "3a0611f76b7e02b393d422a5598e5afd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Produit": {
                "required": [
                    "id_categorie",
                    "id_marque",
                    "nom",
                    "description",
                    "prix",
                    "stock",
                    "image"
                ],
                "properties": {
                    "id_produit": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_categorie": {
                        "type": "integer",
                        "example": 2
                    },
                    "id_promotion": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "id_marque": {
                        "type": "integer",
                        "example": 3
                    },
                    "nom": {
                        "type": "string",
                        "example": "Panneau Solaire X"
                    },
                    "description": {
                        "type": "string",
                        "example": "Haute performance"
                    },
                    "prix": {
                        "type": "number",
                        "format": "float",
                        "example": 499
                    },
                    "stock": {
                        "type": "integer",
                        "example": 20
                    },
                    "image": {
                        "type": "string",
                        "example": "/images/p1.jpg"
                    }
                },
                "type": "object"
            },
...
            "Marque": {
                "required": [
                    "nom",
                    "paysOrigine"
                ],
                "properties": {
                    "id_marque": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "type": "string"
                    },
                    "paysOrigine": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Paiement": {
                "required": [
                    "montant",
                    "modePaiement",
                    "datePaiement"
                ],
                "properties": {
                    "id_paiement": {
                        "type": "integer",
                        "example": 1
                    },
                    "montant": {
                        "type": "number"
                    },
                    "modePaiement": {
                        "type": "string"
                    },
                    "datePaiement": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    }
}